initial refactor -- draft
